[FEAT] #103 Logout Function Add
