+++ src/pages/auth/SignUpPage.vue
@@ -10,7 +10,7 @@ onMounted(() => {
   // 렌더링 0.5초 (500ms → 테스트로 5000ms 설정)
   setTimeout(() => {
     isReady.value = true
-  }, 20000) // ✅ 여긴 원하는 시간으로 조절 가능 (보통 300~800ms 추천)
+  }, 5000) // ✅ 여긴 원하는 시간으로 조절 가능 (보통 300~800ms 추천)
 })
 </script>
